requirements.txt
numpy
simpleaudio
scipy
pyqtgraph
pyserial
